logger and notifications

+++ src/modules/category/category.controller.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from "express";
 import { Res } from "../../utils/Response";
+import { getServicesByInterestedCategories } from "../services/services.service";
 import { CreateCategoryDto, UpdateCategoryDto } from "./category.schema";
 import * as categoryService from "./category.service";
 
@@ -26,7 +27,7 @@ export const createCategory = async (req: Request, res: Response) => {
     }
     const newCategory = await categoryService.createCategory(newCategoryData);
     res.status(201).json(new Res(newCategory, "Categoría creada con éxito"));
-  } catch (error) {
+  } catch {
     res.status(400).json(new Res(null, "Error al crear la categoría", false));
   }
 };
@@ -41,7 +42,7 @@ export const updateCategory = async (req: Request, res: Response) => {
     }
     const newCategory = await categoryService.updateCategory(req.params.id, newCategoryData);
     res.status(201).json(new Res(newCategory, "Categoría actualizada con éxito"));
-  } catch (error) {
+  } catch {
     res.status(400).json(new Res(null, "Error al actualizar la categoría", false));
   }
 };
@@ -51,7 +52,7 @@ export const activeCategoryToggle = async (req: Request, res: Response) => {
   try {
     await categoryService.activeCategoryToggle(categoryId);
     res.status(200).json(new Res(null, "Actividad de la categoría actualizada con éxito"));
-  } catch (error) {
+  } catch {
     res.status(404).json(new Res(null, "Error al actualizar la actividad de la categoría", false));
   }
 };
@@ -59,9 +60,15 @@ export const activeCategoryToggle = async (req: Request, res: Response) => {
 export const deleteCategory = async (req: Request, res: Response) => {
   const categoryId = req.params.id;
   try {
+    const isCategoryInUse = await getServicesByInterestedCategories([categoryId]);
+    if (isCategoryInUse.length) {
+      return res
+        .status(200)
+        .json(new Res(null, "La categoría está en uso y no se puede eliminar", false));
+    }
     await categoryService.deleteCategory(categoryId);
     res.status(200).json(new Res(null, "Categoría eliminada con éxito"));
-  } catch (error) {
+  } catch {
     res.status(404).json(new Res(null, "Error al eliminar la categoría", false));
   }
 };
